implementando JogosAdapter.kt, começando por JogosViewHolder (utilizando viewBinding)

+++ app/src/main/java/com/app/model/Jogo.kt
@@ -1,8 +1,8 @@
 package com.app.model
 
 data class Jogo(
-    var codigo:Int,
-    var nomeJogo:String,
-    var notaJogo:Int,
-    var console:String
+    var codigo: Int,
+    var nomeJogo: String,
+    var notaJogo: Float,
+    var console: String
 )
